chore(swr): hook folder
