+++ pages/class.tsx
@@ -8,7 +8,7 @@ import { EnrollmentList } from "@components/EnrollmentList";
 import { VolunteeringList } from "@components/VolunteeringList";
 import { roles } from ".prisma/client";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import useMe from "@utils/useMe";
+import useMe from "@utils/hooks/useMe";
 import { Loading } from "@components/Loading";
 import { useRouter } from "next/router";
