+++ pages/parent/signup.tsx
@@ -12,7 +12,7 @@ import dynamic from "next/dynamic";
 import { useState } from "react";
 import { GetServerSideProps } from "next"; // Get server side props
 import { getSession } from "next-auth/client";
-import useLocalStorage from "@utils/useLocalStorage";
+import useLocalStorage from "@utils/hooks/useLocalStorage";
 import { ParentInput, roles, locale, province } from "@models/User";
 import colourTheme from "@styles/colours";
 import { mutate } from "swr";
