+++ utils/hooks/useParentRegistration.ts
@@ -1,8 +1,8 @@
 import { EnrollmentCardInfo } from "@models/Enroll";
 import { locale } from "@prisma/client";
 import useSWR from "swr";
-import CardInfoUtil from "./cardInfoUtil";
-import { fetcher } from "./fetcher";
+import CardInfoUtil from "../cardInfoUtil";
+import { fetcher } from "../fetcher";
 
 export type UseParentRegistrationsResponse = {
     enrollments: EnrollmentCardInfo[];
