+++ utils/hooks/useStripeSession.ts
@@ -1,6 +1,6 @@
 import { Stripe } from "services/stripe";
 import useSWR from "swr";
-import { fetcher } from "./fetcher";
+import { fetcher } from "../fetcher";
 
 export type UseStripeSessionResponse = {
     stripeSession: Stripe.Checkout.Session;
